pyproject.toml
[project]
name = "mariliabot"
version = "0.1.0"
description = "Inteligência imobiliária para Marília-SP"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "supabase>=2.0",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
mariliabot = "src.main:main"
